add validation when try to save snapshot to dataverse without saving parameters. and add moscow into datamigration and doc layout in setting. and fix snapshot saving flow issues

+++ seerDiscoverDataTableROMPCF/src/Utils/EstimateAveRate/customization.design.utils.ts
@@ -70,7 +70,7 @@ export const generateCustomisationDesignMValue = async(inititlaData: any, condit
     const {BaseData, resourceModelData, ModuleData, parameterModel, CustomisationModels} = inititlaData
     let {hourlyRate, hoursPerday} = parameterModel[0]
     if (hasParameters) {
-      hoursPerday = parseInt(settingParameters?.formattedData[
+      hoursPerday = parseFloat(settingParameters?.formattedData[
         parameterKeyIndex.hoursPerDay
       ]?.currentValue || '0');
     }
